+++ app/src/main/java/illyan/jay/domain/model/DomainSession.kt
@@ -3,8 +3,8 @@ package illyan.jay.domain.model
 import java.util.*
 
 data class DomainSession(
-    val id: Int = -1,
-    val startTime: Date,
-    var endTime: Date?,
-    var distance: Double = 0.0
+	val id: Long = -1,
+	val startTime: Date,
+	var endTime: Date?,
+	var distance: Double = 0.0
 )
