+++ app/src/main/java/illyan/jay/ui/sessions/session_info/model/UiSession.kt
@@ -3,8 +3,8 @@ package illyan.jay.ui.sessions.session_info.model
 import java.util.*
 
 data class UiSession(
-    val id: Int = -1,
-    val startTime: Date,
-    var endTime: Date? = null, // in millis
-    var distance: Double = 0.0
+	val id: Long = -1,
+	val startTime: Date,
+	var endTime: Date? = null, // in millis
+	var distance: Double = 0.0
 )
